+++ test/conformance/kuttl/generate/policy/standard/data/sync/pol-data-sync-delete-rule/06-checks.yaml
@@ -2,6 +2,6 @@ apiVersion: kuttl.dev/v1beta1
 kind: TestStep
 apply:
 assert:
-- secret.yaml
+- configmap-remain.yaml
 error:
 - configmap.yaml
