+++ test/conformance/kuttl/generate/policy/standard/data/sync/pol-data-sync-delete-rule/configmap-remain.yaml
@@ -1,0 +1,9 @@
+apiVersion: v1
+data:
+  key: superconfigmap
+kind: ConfigMap
+metadata:
+  labels:
+    somekey: somevalue
+  name: superconfigmap
+  namespace: pol-data-sync-delete-rule
