--- test/conformance/kuttl/generate/policy/standard/data/sync/pol-data-sync-delete-rule/secret.yaml
@@ -1,10 +1,0 @@
-apiVersion: v1
-data:
-  mysupersecretkey: bXlzdXBlcnNlY3JldHZhbHVl
-kind: Secret
-metadata:
-  labels:
-    somekey: somesecretvalue
-  name: supersecret
-  namespace: pol-data-sync-delete-rule
-type: Opaque
