+++ test/conformance/kuttl/generate/validation/clusterpolicy/prevent-loop/01-fail.yaml
@@ -1,0 +1,5 @@
+apiVersion: kuttl.dev/v1beta1
+kind: TestStep
+apply:
+- file: policy.yaml
+  shouldFail: true
