+++ test/conformance/kuttl/generate/validation/policy/immutable-downstream/01-policy.yaml
@@ -11,7 +11,7 @@ spec:
       any:
       - resources:
           kinds:
-          - ConfigMap
+          - Secret
     generate:
       synchronize: true
       apiVersion: v1
